Add files via upload
Adding functionality to determine the winner among two players.  Functions determine if a certain poker combination is true, and if so, returning the cards formatted in order of their combinations and rank to help determine a winner.  Straight, Flush, and Straight flush still need to return lists.  everything else appears to be functional